refactor(compute): extract remote operation owner
